+++ packages/i18n/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./config";
+export * from "./components";
 export * from "./hooks";
