+++ packages/types/src/users.d.ts
@@ -25,7 +25,6 @@ export interface IUser extends IUserLite {
   is_password_autoset: boolean;
   is_tour_completed: boolean;
   mobile_number: string | null;
-  role: string | null;
   last_workspace_id: string;
   user_timezone: string;
   username: string;
@@ -62,6 +61,7 @@ export type TUserProfile = {
   billing_address_country: string | undefined;
   billing_address: string | undefined;
   has_billing_address: boolean;
+  language: string;
   created_at: Date | string;
   updated_at: Date | string;
 };
