+++ web/core/components/profile/index.ts
@@ -3,5 +3,6 @@ export * from "./overview";
 export * from "./profile-issues-filter";
 export * from "./sidebar";
 export * from "./time";
-export * from "./profile-setting-content-wrapper"
-export * from "./profile-setting-content-header"
+export * from "./profile-setting-content-wrapper";
+export * from "./profile-setting-content-header";
+export * from "./form";
