+++ web/core/lib/wrappers/store-wrapper.tsx
@@ -1,7 +1,8 @@
-import { ReactNode, useEffect, FC, useState } from "react";
+import { ReactNode, useEffect, FC } from "react";
 import { observer } from "mobx-react";
 import { useParams } from "next/navigation";
 import { useTheme } from "next-themes";
+import { useTranslation, Language } from "@plane/i18n";
 // helpers
 import { applyTheme, unsetCustomCssVariables } from "@/helpers/theme.helper";
 // hooks
@@ -21,6 +22,7 @@ const StoreWrapper: FC<TStoreWrapper> = observer((props) => {
   const { setQuery } = useRouterParams();
   const { sidebarCollapsed, toggleSidebar } = useAppTheme();
   const { data: userProfile } = useUserProfile();
+  const { changeLanguage } = useTranslation();
 
   /**
    * Sidebar collapsed fetching from local storage
@@ -28,7 +30,6 @@ const StoreWrapper: FC<TStoreWrapper> = observer((props) => {
   useEffect(() => {
     const localValue = localStorage && localStorage.getItem("app_sidebar_collapsed");
     const localBoolValue = localValue ? (localValue === "true" ? true : false) : false;
-
     if (localValue && sidebarCollapsed === undefined) toggleSidebar(localBoolValue);
   }, [sidebarCollapsed, setTheme, toggleSidebar]);
 
@@ -37,7 +38,6 @@ const StoreWrapper: FC<TStoreWrapper> = observer((props) => {
    */
   useEffect(() => {
     if (!userProfile?.theme?.theme) return;
-
     const currentTheme = userProfile?.theme?.theme || "system";
     const currentThemePalette = userProfile?.theme?.palette;
     if (currentTheme) {
@@ -50,6 +50,11 @@ const StoreWrapper: FC<TStoreWrapper> = observer((props) => {
       } else unsetCustomCssVariables();
     }
   }, [userProfile?.theme?.theme, userProfile?.theme?.palette, setTheme]);
+
+  useEffect(() => {
+    if (!userProfile?.language) return;
+    changeLanguage(userProfile?.language as Language);
+  }, [userProfile?.language, changeLanguage]);
 
   useEffect(() => {
     if (!params) return;
